Refactor Blueprint parsing
This is a very weird refactor, I did mostly to try using sealed classes for it, but doesn't seem
to be working that well. I will probably come back and try something else. like having CreateFolder
with a nullable type for the children, maybe that would allow better use of Jackson.